Split Examples to Test Files
Removed examples_test.go file and moved the testable example functions into polo_test.go and document_test.go
